+++ tests/EntityManager.sqlite2.test.ts
@@ -14,6 +14,8 @@ import { initORMSqlite2, mockLogger } from './bootstrap';
 import type { IAuthor4, IPublisher4, ITest4 } from './entities-schema';
 import { Author4, Book4, BookTag4, FooBar4, Publisher4, PublisherType, Test4 } from './entities-schema';
 
+jest.retryTimes(3);
+
 describe.each(['sqlite', 'better-sqlite'] as const)('EntityManager (%s)', driver => {
 
   let orm: MikroORM;