+++ tests/setup.ts
@@ -5,4 +5,5 @@ process.env.MIKRO_ORM_ALLOW_GLOBAL_CLI = '1';
 process.env.MIKRO_ORM_ALLOW_VERSION_MISMATCH = '1';
 
 jest.restoreAllMocks();
+jest.retryTimes(1);
 MetadataStorage.clear();